requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
freezegun==1.2.2
inflection==0.5.1
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
voluptuous==0.13.1
